LabAssistantMobileApp/requirements.txt
Flask
google-genai
Pillow
gunicorn
